chore: resolved issues for PR #81
