+++ backend/src/config/app-config.config.ts
@@ -1,4 +1,5 @@
 export const ALLOWED_ORIGINS = {
   TESTNET: ["http://localhost:8080", "https://cdpn.io"],
-  MAINNET: []
+  MAINNET: [],
+  DEV: []
 };
