+++ backend/src/routes/publications.route.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { getRelatedPublicationsController } from "../controllers/publications.controller";
-import { validateRelatedRouteRequestMiddleware } from "../middlewares/validate-request.publications.middleware";
+import { validateRelatedRouteRequestMiddleware } from "../middlewares/publications/validate-request.publications.middleware";
 
 const router = express.Router();
