+++ backend/src/utils/publications/get-publications-for-tag.publications.util.ts
@@ -2,6 +2,13 @@ import { getRelatedParentPublicationsUtil } from "../related-parent-publications
 import { FAILURE, SUCCESS } from "../../config/app-constants.config";
 import { PostPublicationsLensProtocolModel } from "../../models/lens-protocol/post.publications.lens-protocol.model";
 
+/**
+ * Finds publications related to the input string.
+ *
+ * @param {string} inputString - The input string to search for related publications.
+ * @return {Promise<{ relatedPublications: string[], message: string }>} - An object containing an array
+ *   of related publication IDs and a success/failure message.
+ */
 export const getPublicationsForTagPublicationsUtil = async (
   inputString: string
 ) => {
